+++ src/model/training.py
@@ -1,14 +1,168 @@
-"""This module contains methods to train our PyTorch model"""
-
+from tempfile import TemporaryDirectory
+from typing import Any
+import time
 import torch
+from pathlib import Path
+
+# ! REMOVE
+dataset_sizes = {"train": 51058, "val": 14959, "test": 1793}
 
 
-def fine_tune(
-    model: torch.nn,
-    loss: torch.nn,
+def __train_epoch(
+    model: torch.nn.Module,
+    dataloader: torch.utils.data.DataLoader,
+    criterion: torch.nn.Module,
     optimizer: torch.optim,
-    epochs: int,
-    lr_scheduler: torch.lr_scheduler,
-    head: bool = True,
-):
-    pass
+    device: Any,
+) -> tuple[float, int]:
+    """Trains a PyTorch model for one epoch
+
+    Args:
+        model (torch.nn.Module): The model to train
+        dataloader (torch.utils.data.DataLoader): The dataloader for the test set
+        criterion (torch.nn.Module): The loss function
+        optimizer (torch.optim): The optimization function
+        device (Any): The device to compute on
+
+    Returns:
+        tuple[float, int]: The running loss and running number of correct predctions
+    """
+    model.train()  # Set model to training mode
+    running_loss: float = 0.0
+    running_corrects: int = 0
+
+    # Iterate over data.
+    for inputs, labels in dataloader:
+        inputs: torch.Tensor = inputs.to(device)
+        labels: torch.Tensor = labels.to(device)
+
+        # zero the parameter gradients
+        optimizer.zero_grad()
+
+        # forward
+        outputs = model(inputs)
+        _, preds = torch.max(outputs, 1)  # ? why do we need this?
+        loss = criterion(outputs, labels)
+
+        # backward + optimize
+        loss.backward()
+        optimizer.step()
+
+        # statistics
+        running_loss += loss.item() * inputs.size(0)
+        running_corrects += torch.sum(preds == labels.data)
+
+    return running_loss, running_corrects
+
+
+def __validate_epoch(
+    model: torch.nn.Module,
+    dataloader: torch.utils.data.DataLoader,
+    criterion: torch.nn.Module,
+    device: Any,
+) -> tuple[float, int]:
+    """Validates a PyTorch model for one epoch
+
+    Args:
+        model (torch.nn.Module): The model to train
+        dataloader (torch.utils.data.DataLoader): The dataloader for the validation set
+        criterion (torch.nn.Module): The loss function
+        device (Any): The device to compute on
+
+    Returns:
+        tuple[float, int]: The running loss and running number of correct predctions
+    """
+    model.eval()  # Set model to evaluate mode
+    running_loss = 0.0
+    running_corrects = 0
+
+    # Iterate over data.
+    for inputs, labels in dataloader:
+        inputs = inputs.to(device)
+        labels = labels.to(device)
+
+        # forward
+        with torch.no_grad():
+            outputs = model(inputs)
+            _, preds = torch.max(outputs, 1)
+            loss = criterion(outputs, labels)
+
+        # statistics
+        running_loss += loss.item() * inputs.size(0)
+        running_corrects += torch.sum(preds == labels.data)
+
+    return running_loss, running_corrects
+
+
+def train_model(
+    model: torch.nn.Module,
+    dataloaders: dict[str, torch.utils.data.DataLoader],
+    criterion: torch.nn.Module,
+    optimizer: torch.optim,
+    scheduler: torch.optim.lr_scheduler,
+    device: Any,
+    num_epochs: int = 3,
+) -> tuple[torch.nn.Module, list]:
+    """Trains a PyTorch model for the number of specified epochs
+
+    Args:
+        model (torch.nn.Module): The model to train
+        dataloaders (dict[str, torch.utils.data.DataLoader]): The dataloaders containing our training data
+        criterion (torch.nn.Module): The loss function
+        optimizer (torch.optim): The optimization function
+        scheduler (torch.optim.lr_scheduler): The learning rate scheduler
+        device (Any): The device to do the training on
+        num_epochs (int, optional): The number of epochs to train. Defaults to 3.
+
+    Returns:
+        tuple[torch.nn.Module, list]: The best model and training history
+    """
+    since: float = time.time()
+
+    model.to(device)
+
+    # Create a temporary directory to save training checkpoints
+    with TemporaryDirectory() as tempdir:
+        best_model_params_path: Path = Path(tempdir) / "best_model_params.pt"
+        torch.save(model.state_dict(), best_model_params_path)
+        best_acc: float = 0.0
+        history: list = []
+
+        for epoch in range(num_epochs):
+            # Each epoch has a training and validation phase
+            train_loss, train_corrects = __train_epoch(
+                model, dataloaders["train"], criterion, optimizer, device
+            )
+            if scheduler:
+                scheduler.step()
+            val_loss, val_corrects = __validate_epoch(
+                model, dataloaders["val"], criterion, device
+            )
+
+            train_loss /= dataset_sizes["train"]
+            train_acc = train_corrects.double() / dataset_sizes["train"]
+            val_loss /= dataset_sizes["val"]
+            val_acc = val_corrects.double() / dataset_sizes["val"]
+
+            history.append([train_acc, val_acc, train_loss, val_loss])
+            print(
+                f"Epoch {epoch}/{num_epochs - 1}: "
+                f"Train Loss: {train_loss:.4f}, Acc: {train_acc:.4f}, "
+                f"Val Loss: {val_loss:.4f}, Acc: {val_acc:.4f}"
+            )
+
+            # deep copy the model
+            if val_acc > best_acc:
+                best_acc = val_acc
+                torch.save(model.state_dict(), best_model_params_path)
+
+        time_elapsed: float = time.time() - since
+        print(
+            f"\nTraining complete in {time_elapsed // 60:.0f}m {time_elapsed % 60:.0f}s"
+        )
+        print(f"Best val Acc: {best_acc:.4f}")
+
+        # load best model weights
+        model.load_state_dict(torch.load(best_model_params_path))
+
+    return model, history