+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Adds initial project documents
 - Adds a `HiRISE` dataset loader class and associated unit tests
+- [1-2-2024]
+    - Adds the model/adapt.py module to modify pre-trained models for HiRISE
 
 ### Changed